1.修改之前遗留Bug  2.删除角色整理完成

+++ Unity/Codes/Model/Demo/Role/RoleInfosComponent.cs
@@ -10,6 +10,6 @@ namespace ET
     public class RoleInfosComponent : Entity,IAwake,IDestroy
     {
         public List<RoleInfo> RoleInfos = new List<RoleInfo>();
-        public int CurrentRoleId = 0;
+        public long CurrentRoleId = 0;
     }
 }
